requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.2
dataclasses-json==0.5.13
diskcache==5.6.1
docx2txt==0.8
et-xmlfile==1.1.0
faiss-cpu==1.7.4
filelock==3.12.2
filetype==1.2.0
frozenlist==1.4.0
fsspec==2023.6.0
gpt4all==0.1.7
huggingface-hub==0.16.4
idna==3.4
joblib==1.3.1
langchain==0.0.242
langsmith==0.0.14
llama-cpp-python==0.1.77
lxml==4.9.3
Markdown==3.4.4
marshmallow==3.20.1
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.1
olefile==0.46
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==10.0.0
pycparser==2.21
pydantic==1.10.12
pypandoc==1.11
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
six==1.16.0
SQLAlchemy==2.0.19
tabulate==0.9.0
tenacity==8.2.2
tokenizers==0.13.3
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
unstructured==0.8.1
urllib3==2.0.4
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
